### 2023-03-09 #5237 @editorialbot
[PRE REVIEW]: ndcube: Manipulating N-dimensional Astronomical Data in Python
**Submitting author:** @DanRyanIrish (<a href="http://orcid.org/0000-0001-8661-3825">Daniel Ryan</a>)
**Repository:** https://github.com/sunpy/ndcube
**Branch with paper.md** (empty if default branch):
**Version:** v2.1.1
