### 2023-03-09 6425f7c
Add author to JOSS paper.
