+++ joss_paper/paper.md
@@ -38,6 +38,9 @@ authors:
   - name: J. Marcus Hughes
     orcid: 0000-0003-3410-7650
     affiliation: 14
+  - name: David Stansby
+    orcid: 0000-0002-1365-1908
+    affiliation: 15
   - name: Albert Y. Shih
     orcid: 0000-0001-6874-2594
     affiliation: 12
@@ -73,6 +76,8 @@ affiliations:
    index: 13
  - name: Southwest Research Institute, USA
    index: 14
+ - name: Advanced Research Computing Centre, University College London, UK
+   index: 15
 date: 24 February 2023
 bibliography: paper.bib
 
