### 2023-03-10 #5237 @DanRyanIrish
> @DanRyanIrish — Thanks for your submission! I'm working to assign an editor to this issue and they'll take it from there.
>
> I believe that this is a [joint submission with the AAS journals](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing) - is that right, @DanRyanIrish? When you have the DOI for your AAS paper, we need to remember to add the following fields to your JOSS manuscript:
>
> ```yaml
> aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
> aas-journal: Astrophysical Journal <- The name of the AAS journal.
> ```

Hi @dfm.  Thanks for managing this submission.  Yes that's right.  This is a joint submission with AAS.  As soon as I have a DOI I will let you know.
