### 2023-03-10 #5237 @DanRyanIrish
@dfm : I noticed that one author is missing from the proof, David Stansby.  His name is in the paper.md doc in the ndcube repo so hopefully the fix is to simply regenerate a proof from that file?  Many thanks.
