### 2023-03-11 c1772ed
Add nitpicky, reformat docs filepath and spell check
