+++ ndcube/ndcube.py
@@ -83,8 +83,8 @@ class NDCubeABC(astropy.nddata.NDDataBase):
         multiple array axes, the same physical type string can appear in multiple tuples.
 
         The physical types returned by this property are drawn from the
-        `~NDCube.combined_wcs` property so they include the coordinates contained in
-        `~NDCube.extra_coords`.
+        `~ndcube.NDCube.combined_wcs` property so they include the coordinates contained in
+        `~ndcube.NDCube.extra_coords`.
         """
 
     @abc.abstractmethod
@@ -100,7 +100,7 @@ class NDCubeABC(astropy.nddata.NDDataBase):
         ----------
         axes: `int` or `str`, or multiple `int` or `str`, optional
             Axis number in numpy ordering or unique substring of
-            `~ndcube.NDCube.world_axis_physical_types`
+            `ndcube.NDCube.wcs.world_axis_physical_types <astropy.wcs.wcsapi.BaseWCSWrapper>`
             of axes for which real world coordinates are desired.
             Not specifying axes inputs causes results for all axes to be returned.
         pixel_corners: `bool`, optional
@@ -139,14 +139,15 @@ class NDCubeABC(astropy.nddata.NDDataBase):
                                  ) -> Iterable[u.Quantity]:
         """
         Returns the world coordinate values of all pixels for desired axes.
-        In contrast to `axis_world_coords()`, this method returns `~astropy.units.Quantity` objects. which only
-        provide units rather than full coordinate metadata provided by high-level coordinate objects.
+        In contrast to :meth:`ndcube.NDCube.axis_world_coords`, this method returns
+        `~astropy.units.Quantity` objects. Which only provide units rather than full
+        coordinate metadata provided by high-level coordinate objects.
 
         Parameters
         ----------
         axes: `int` or `str`, or multiple `int` or `str`, optional
             Axis number in numpy ordering or unique substring of
-            `~ndcube.NDCube.wcs.world_axis_physical_types`
+            `ndcube.NDCube.wcs.world_axis_physical_types <astropy.wcs.wcsapi.BaseWCSWrapper>`
             of axes for which real world coordinates are desired.
             axes=None implies all axes will be returned.
 
@@ -212,7 +213,7 @@ class NDCubeABC(astropy.nddata.NDDataBase):
 
         Returns
         -------
-        result: `~ndcube..ndcube.NDCubeABC`
+        result: `~ndcube.ndcube.NDCubeABC`
 
         Examples
         --------
@@ -235,9 +236,9 @@ class NDCubeABC(astropy.nddata.NDDataBase):
         contains all the provided world coordinate points.
 
         This function takes points as iterables of low-level coordinate objects,
-        i.e. `~astropy.units.Quantity` objects.  This differs from `~ndcube.NDCube.crop()`
+        i.e. `~astropy.units.Quantity` objects. This differs from `~ndcube.NDCube.crop()`
         which takes high-level coordinate objects requiring all the relevant coordinate
-        information such as coordinate frame etc.  Hence this method's API is more basic
+        information such as coordinate frame etc. Hence this method's API is more basic
         but less explicit.
 
 
@@ -265,12 +266,10 @@ class NDCubeABC(astropy.nddata.NDDataBase):
         -------
         result: `~ndcube.ndcube.NDCubeABC`
 
-
         Examples
         --------
         An example of cropping a region of interest on the Sun from a 3-D image-time cube:
         >>> NDCube.crop_by_values((-600, -600, 0), (0, 0, 0), units=(u.arcsec, u.arcsec, u.s)) # doctest: +SKIP
-
         """
 
 
@@ -324,23 +323,23 @@ class NDCubeBase(NDCubeABC, astropy.nddata.NDData, NDCubeSlicingMixin):
 
     Parameters
     ----------
[…]
