+++ ndcube/extra_coords/__init__.py
@@ -1,5 +1,6 @@
 from .extra_coords import ExtraCoords, ExtraCoordsABC
-from .table_coord import *
+from .table_coord import (BaseTableCoordinate, MultipleTableCoordinate, QuantityTableCoordinate,
+                          SkyCoordTableCoordinate, TimeTableCoordinate)
 
-__all__ = ['ExtraCoordsABC', 'ExtraCoords', 'TimeTableCoordinate',
-           'SkyCoordTableCoordinate', 'QuantityTableCoordinate']
+__all__ = ['ExtraCoordsABC', 'ExtraCoords', 'TimeTableCoordinate', "MultipleTableCoordinate",
+           'SkyCoordTableCoordinate', 'QuantityTableCoordinate', "BaseTableCoordinate"]

[…]
