+++ ndcube/extra_coords/__init__.py
@@ -1,6 +1,5 @@
-from .extra_coords import ExtraCoords, ExtraCoordsABC
 from .table_coord import (BaseTableCoordinate, MultipleTableCoordinate, QuantityTableCoordinate,
                           SkyCoordTableCoordinate, TimeTableCoordinate)
 
-__all__ = ['ExtraCoordsABC', 'ExtraCoords', 'TimeTableCoordinate', "MultipleTableCoordinate",
+__all__ = ['TimeTableCoordinate', "MultipleTableCoordinate",
            'SkyCoordTableCoordinate', 'QuantityTableCoordinate', "BaseTableCoordinate"]
