+++ ndcube/ndcube_sequence.py
@@ -8,7 +8,7 @@ import numpy as np
 from ndcube import utils
 from ndcube.visualization.descriptor import PlotterDescriptor
 
-__all__ = ['NDCubeSequence']
+__all__ = []
 
 
 class NDCubeSequenceBase:
