### 2023-03-15 1e3c6ca
minor tweaks for specutils
