+++ ndcube/visualization/base.py
@@ -1,7 +1,5 @@
 import abc
 
-__all__ = ['BasePlotter']
-
 
 class BasePlotter(abc.ABC):
     """
