### 2023-03-23 #5296 @editorialbot
[REVIEW]: ndcube: Manipulating N-dimensional Astronomical Data in Python
**Submitting author:** @DanRyanIrish (<a href="http://orcid.org/0000-0001-8661-3825">Daniel Ryan</a>)
**Repository:** https://github.com/sunpy/ndcube
**Branch with paper.md** (empty if default branch):
**Version:** v2.1.3
