### 2023-03-24 #5296 @timj
I assume I can leave a general comment before I check all the boxes since checking boxes seems quite final.

The ndcube looks fantastic and I'm really happy that it is based off NDData and can use gwcs. I think I would like a little more description of the state of the field to be included since it all seems to be wrapped up in a couple of sentences.

I would like a more explicit statement about support for gWCS. When I first read the paper I assumed FITS-WCS was all that was supported and it was only after reading the ndcube docs that I realized that there is full support for gWCS. This is really important since FITS-WCS is very restrictive in what can be supported. For example, one party trick the Starlink AST library has is that you can define an IFU 3D coordinate system on a 2D image (this can not be handled in FITS-WCS at all since it requires discrete jumps in cube planes moving from one pixel to the next) and ask AST to regrid it into a 3D cube and it will do it more or less automatically. Can ndcube do that?

References:

* The jdaviz reference has a broken author since it thinks that it's done by "Developers, J.".
* Can we include the URL to the CuPy conference paper? I think it's http://learningsys.org/nips17/assets/papers/paper_16.pdf
* Would it be possible for the ApJ paper to be uploaded to arXiv so it can get a DOI that can be included here? Does JOSS allow "submitted" papers that may or may not happen?

Typo on bottom of first page: "is treats them"
