### 2023-03-24 #5296 @timj
Also there is a typo in the name of the package in the main readme ("ncdube"):

https://github.com/sunpy/ndcube/blob/main/README.rst?plain=1#L2
