### 2023-03-27 #5296 @DanRyanIrish
Thanks @timj for these comments.  I will open a PR to the ndcube joss_paper branch so you can see how I've tried to address your comments.  If there's a preferred way of working through these comments, please let me know.
