### 2023-03-31 1e99044
Some minor corrections from JOSS review.
