+++ joss_paper/paper.bib
@@ -27,6 +27,7 @@
   booktitle = {Proceedings of Workshop on Machine Learning Systems ({{LearningSys}}) in the Thirty-First Annual Conference on Neural Information Processing Systems ({{NIPS}})},
   author = {Okuta, Ryosuke and Unno, Yuya and Nishino, Daisuke and Hido, Shohei and Loomis, Crissman},
   year = {2017}
+  url = {http://learningsys.org/nips17/assets/papers/paper_16.pdf}
 }
 
 @Manual{dask,
@@ -91,7 +92,7 @@
 }
 
 @software{jdaviz,
-  author       = {JDADF Developers and
+  author       = {JDADF-Developers and
                   Averbukh, Jesse and
                   Bradley, Larry and
                   Buikhuizen, Mario and
