### 2023-03-31 #5296 @DanRyanIrish
@timj I have opened [this PR](https://github.com/sunpy/ndcube/pull/609/files) to the ndcube main branch with changes in accordance with your comments.  This is still a work in progress and I'll let you know when I think it's ready for further feedback.  But in the mean time you can see what I'm doing.
