### 2023-03-31 #5296 @DanRyanIrish
> ...For example, one party trick the Starlink AST library has is that you can define an IFU 3D coordinate system on a 2D image (this can not be handled in FITS-WCS at all since it requires discrete jumps in cube planes moving from one pixel to the next) and ask AST to regrid it into a 3D cube and it will do it more or less automatically. Can ndcube do that?

I think @Cadair may want to comment here.  From my understanding of your question, `NDCube` does not explicitly support this.  `NDCube` does have `.reproject_to` method that takes a new WCS and reprojects the cube to it.  The actual calculation is passed onto the `reproject` package.  So if what you're asking for can already be done by `reproject`, then `NDCube` already supports it.  If not however, there is no reason that an `NDCube`-inherited object couldn't add a method that does provide such functionality.
