### 2023-03-31 #5296 @DanRyanIrish
>     * Would it be possible for the ApJ paper to be uploaded to arXiv so it can get a DOI that can be included here? Does JOSS allow "submitted" papers that may or may not happen?

I don't know JOSS's policy on this.  Perhaps @adonath can clarify?
