### 2023-03-31 ad05ed5
Explicitly mention in JOSS paper support for gWCS and make some other points a bit clearer.
