+++ joss_paper/paper.bib
@@ -198,9 +198,37 @@ Transformations in Python: The NDCube 2 & Astropy APE-14 WCS APIs}",
   url = {https://github.com/astropy/specutils}
 }
 
+@ARTICLE{sunpy,
+       author = {{Mumford}, Stuart and {Freij}, Nabil and {Christe}, Steven and {Ireland}, Jack and {Mayer}, Florian and {Hughitt}, V. and {Shih}, Albert and {Ryan}, Daniel and {Liedtke}, Simon and {P{\'e}rez-Su{\'a}rez}, David and {Chakraborty}, Pritish and {K}, Vishnunarayan and {Inglis}, Andrew and {Pattnaik}, Punyaslok and {Sip{\H{o}}cz}, Brigitta and {Sharma}, Rishabh and {Leonard}, Andrew and {Stansby}, David and {Hewett}, Russell and {Hamilton}, Alex and {Hayes}, Laura and {Panda}, Asish and {Earnshaw}, Matt and {Choudhary}, Nitin and {Kumar}, Ankit and {Chanda}, Prateek and {Haque}, Md and {Kirk}, Michael and {Mueller}, Michael and {Konge}, Sudarshan and {Srivastava}, Rajul and {Jain}, Yash and {Bennett}, Samuel and {Baruah}, Ankit and {Barnes}, Will and {Charlton}, Michael and {Maloney}, Shane and {Chorley}, Nicky and {Himanshu} and {Modi}, Sanskar and {Mason}, James and {Naman} and {Campos Rozo}, Jose Ivan and {Manley}, Larry and {Chatterjee}, Agneet and {Evans}, John and {Malocha}, Michael and {Bobra}, Monica and {Ghosh}, Sourav and {Airmansmith} and {Sta{\'n}czak}, Dominik and {De Visscher}, Ruben and {Verma}, Shresth and {Agrawal}, Ankit and {Buddhika}, Dumindu and {Sharma}, Swapnil and {Park}, Jongyeob and {Bates}, Matt and {Goel}, Dhruv and {Taylor}, Garrison and {Cetusic}, Goran and {Jacob} and {Inchaurrandieta}, Mateo and {Dacie}, Sally and {Dubey}, Sanjeev and {Sharma}, Deepankar and {Bray}, Erik and {Rideout}, Jai and {Zahniy}, Serge and {Meszaros}, Tomas and {Bose}, Abhigyan and {Chicrala}, Andr{\'e} and {Ankit} and {Guennou}, Chlo{\'e} and {D'Avella}, Daniel and {Williams}, Daniel and {Ballew}, Jordan and {Murphy}, Nick and {Lodha}, Priyank and {Robitaille}, Thomas and {Krishan}, Yash and {Hill}, Andrew and {Eigenbrot}, Arthur and {Mampaey}, Benjamin and {Wiedemann}, Bernhard and {Molina}, Carlos and {Ke{\c{s}}kek}, Duygu and {Habib}, Ishtyaq and {Letts}, Joseph and {Baz{\'a}n}, Juanjo and {Arbolante}, Quinn and {Gomillion}, Reid and {Kothari}, Yash and {Sharma}, Yash and {Stevens}, Abigail and {Price-Whelan}, Adrian and {Mehrotra}, Ambar and {Kustov}, Arseniy and {Stone}, Brandon and {Dang}, Trung and {Arias}, Emmanuel and {Dover}, Fionnlagh and {Verstringe}, Freek and {Kumar}, Gulshan and {Mathur}, Harsh and {Babuschkin}, Igor and {Wimbish}, Jaylen and {Buitrago-Casas}, Juan and {Krishna}, Kalpesh and {Hiware}, Kaustub
[…]
