### 2023-03-31 #5296 @DanRyanIrish
@timj With the exception of deciding how best to handle the reference to the ndcube ApJ paper, I have made changes to address all your comments.  You can this the changes on [the PR mentioned above](https://github.com/sunpy/ndcube/pull/609)

Thanks again for your constructive feedback.
