### 2023-03-31 #5296 @adonath
> I don't know JOSS's policy on this. Perhaps @adonath can clarify?

@DanRyanIrish and @timj The policy by JOSS here is to wait for the AAS DOI. Once the DOI is available it should be added to the YAML header of the `paper.md` file. The guidelines are documented here https://joss.readthedocs.io/en/latest/editing.html?#handling-of-papers-published-together-with-aas-publishing

I'm currently checking with the JOSS editorial board the policy for the hypothetical case that the paper is not accepted by AAS. I will keep you updated once I know more.
