### 2023-04-05 074bfb1
Minor edits to JOSS paper
