### 2023-04-05 5441ec8
More minor corrections to JOSS paper
