+++ joss_paper/paper.md
@@ -93,15 +93,15 @@ Its data classes link data and their coordinates and provide analysis
 methods to manipulate them self-consistently.
 These aim to provide simple and intuitive ways of handling coordinate-aware data,
 analogous to how users handle coordinate-agnostic data with arrays.
-ndcube requires that coordinate transformations be be expressed via the World
+ndcube requires that coordinate transformations be expressed via the World
 Coordinate System (WCS), a coordinate framework commonly used throughout
 astronomy.
 The WCS framework has multiple implementations, e.g. FITS-WCS, gWCS, etc.,
 each with a different incompatible API, which makes workflows and derived tools
 non-transferable between implementations.
-ndcube overcomes this challenge by leveraging AstroPy's WCS API [APE-14; @ape14]
-which can be wrapped around underlying WCS implementations, thereby creating an
-implementation-agnostic API for interacting with the WCS coordinate transformations
+ndcube overcomes this by leveraging AstroPy's WCS API [APE-14; @ape14]
+which can be wrapped around underlying WCS implementations, thereby enabling ndcube
+to support any WCS implementation with the same API.
 ndcube's data-WCS coupling allows users to analyze their data more easily and
 reliably, thus helping to boost their scientific output.
 
