### 2023-04-05 #5296 @DanRyanIrish
@adonath I have received the referee's report on the ApJ paper.  It's very positive so there doesn't appear to be any risk of the paper being rejected.

@timj Are you happy that the changes made in [my ndcube PR](https://github.com/sunpy/ndcube/pull/609) address all your concerns?
