### 2023-04-05 #5296 @Cadair
@timj This probably isn't the best place to go into the gritty details on this, but there's no reason why NDCube couldn't support the Startlink AST WCS library directly via the APE 14 API. I would assume that the regrid would be easier using the startlink lib rather than going through reproject but it might work :grinning:
