### 2023-04-05 sunpy/ndcube#610 @maxnoe
JOSS paper review: typos / wording
### Provide a general description of the issue or problem.

Reviewing the JOSS paper proposal, I found the following simple typos that should be easily fixed:

- [ ] l. 43: a *standarized* framework (missing d in standardized)
- [ ] l. 71: "kwargs" not sure if this is a bit too much "python jargon". Simply write out keyword arguments?
- [ ] l. 141: "facilitate synergies between new combinations of data" This sounds a bit too buzzwordy for me. Maybe: "Makes it easier to combine data from multiple instruments" ?
- [ ] l. 145: missing comma before "the Daniel K. Inouye Solar Telescope" ?
- [ ] l. 149: Wouldn't it be more appropriate to cite the latest astropy paper instead of the first?
- [ ] l 156: Error in authorlist "Developers, J" → "JDAVIZ Developers"
