### 2023-04-05 #5296 @maxnoe
Sorry for being a bit late.

I intentionally didn't read the review comments by @timj , so there might be some duplication.

I found some smaller typos and an issue in the references, for which I opened:
https://github.com/sunpy/ndcube/issues/610

I think this is really great and I am looking forward to try to use ndcube also for high energy gamma-ray data, we also have a lot of use-cases where ndcube might be appropriate.

Side note: the WCS API needs an overhaul to be a bit more user friendly. At least I find myself thinking about one axis at the time and having to specify the properties as lists over the axes instead of just having a list of axis definitions is always a bit counter-intuitive.
