### 2023-04-05 #5296 @DanRyanIrish
Hi @maxnoe.  Thanks for your feedback and for your issue on the ndcube repo.  I will incorporate the necessary changes into the [same PR](https://github.com/sunpy/ndcube/pull/609) I am using to address for @timj 's comments.

> Side note: the WCS API needs an overhaul to be a bit more user friendly. At least I find myself thinking about one axis at the time and having to specify the properties as lists over the axes instead of just having a list of axis definitions is always a bit counter-intuitive.

I don't quite follow here.  Do you mean the AstroPy WCS API?
