### 2023-04-05 #5296 @maxnoe
> I don't quite follow here. Do you mean the AstroPy WCS API?

Yes, nothing directly to do with ndcube, you just use the astropy WCS object
