### 2023-04-05 sunpy/ndcube#610 @DanRyanIrish
These comments will be addressed in #609
