### 2023-04-05 e64aa78
Changes to JOSS paper based on referee comments.
