### 2023-04-05 #5296 @DanRyanIrish
> Yes, nothing directly to do with ndcube, you just use the astropy WCS object

Ah, right.  Now I see why you labelled it a side note.  :)
