### 2023-04-05 #5296 @DanRyanIrish
@timj @maxnoe I have addressed all your comments in sunpy/ndcube#609.  Once the tests pass I will merge the PR and the updates should be available once the editorialbot has regenerated the PDF.
