### 2023-04-05 d37f30d
Add missing comma to bibtex reference in JOSS paper.
