+++ joss_paper/paper.bib
@@ -68,7 +68,7 @@ archivePrefix = {arXiv},
   title = {{{CuPy}}: {{A NumPy-Compatible}} Library for {{NVIDIA GPU}} Calculations},
   booktitle = {Proceedings of Workshop on Machine Learning Systems ({{LearningSys}}) in the Thirty-First Annual Conference on Neural Information Processing Systems ({{NIPS}})},
   author = {Okuta, Ryosuke and Unno, Yuya and Nishino, Daisuke and Hido, Shohei and Loomis, Crissman},
-  year = {2017}
+  year = {2017},
   url = {http://learningsys.org/nips17/assets/papers/paper_16.pdf}
 }
 
