### 2023-04-05 cac3030
Apply suggestions from code review
Co-authored-by: DanRyanIrish <ryand5@tcd.ie>
