+++ setup.cfg
@@ -165,7 +165,7 @@ exclude_lines =
   pragma: py{ignore_python_version}
 
 [codespell]
-skip = *.asdf,*.fits,*.fts,*.header,*.json,*.xsh,*cache*,*egg*,*extern*,.git,.idea,.tox,_build,*truncated,*.svg
+skip = *.asdf,*.fits,*.fts,*.header,*.json,*.xsh,*cache*,*egg*,*extern*,.git,.idea,.tox,_build,*truncated,*.,*svg,*.bib
 ignore-words-list =
     alog,
     nd,
