### 2023-04-17 #5296 @adonath
@timj I think there are a few points left unchecked on your list, but I think you finished your review. I this this correct?

@DanRyanIrish Once the JOSS review is complete, I'll pause the issue and we wait until the AAS paper is published. Please keep us up to date.
