### 2023-05-14 0ad4239
Backport PR #617: Update chat url due to member limit
