### 2023-05-31 d730ea0
Merge pull request #607 from nabobalis/main
Add nitpicky, reformat docs filepath and spell check
