+++ ndcube/ndcube.py
@@ -22,7 +22,7 @@ from astropy.wcs.utils import _split_matrix
 from astropy.wcs.wcsapi import BaseHighLevelWCS, HighLevelWCSWrapper
 
 from ndcube import utils
-from ndcube.extra_coords import ExtraCoords, ExtraCoordsABC
+from ndcube.extra_coords.extra_coords import ExtraCoords, ExtraCoordsABC
 from ndcube.global_coords import GlobalCoords, GlobalCoordsABC
 from ndcube.mixins import NDCubeSlicingMixin
 from ndcube.ndcube_sequence import NDCubeSequence
@@ -30,7 +30,7 @@ from ndcube.utils.wcs_high_level_conversion import values_to_high_level_objects
 from ndcube.visualization import PlotterDescriptor
 from ndcube.wcs.wrappers import CompoundLowLevelWCS, ResampledLowLevelWCS
 
-__all__ = ['NDCubeABC', 'NDCubeBase', 'NDCube']
+__all__ = ['NDCubeABC', 'NDCubeLinkedDescriptor']
 
 # Create mapping to masked array types based on data array type for use in analysis methods.
 ARRAY_MASK_MAP = {}
@@ -66,7 +66,7 @@ class NDCubeABC(astropy.nddata.NDDataBase):
 
         This transform should implement the high level wcsapi, and have
         ``pixel_n_dim`` equal to the number of array dimensions in the
-        `.NDCube`. The number of world dimensions should be equal to the
+        `~ndcube.NDCube`. The number of world dimensions should be equal to the
         number of world dimensions in ``self.wcs`` and in ``self.extra_coords`` combined.
         """
 
@@ -83,8 +83,8 @@ class NDCubeABC(astropy.nddata.NDDataBase):
         multiple array axes, the same physical type string can appear in multiple tuples.
 
         The physical types returned by this property are drawn from the
-        `~NDCube.combined_wcs` property so they include the coordinates contained in
-        `~NDCube.extra_coords`.
+        `~ndcube.NDCube.combined_wcs` property so they include the coordinates contained in
+        `~ndcube.NDCube.extra_coords`.
         """
 
     @abc.abstractmethod
@@ -100,7 +100,7 @@ class NDCubeABC(astropy.nddata.NDDataBase):
         ----------
         axes: `int` or `str`, or multiple `int` or `str`, optional
             Axis number in numpy ordering or unique substring of
-            `~ndcube.NDCube.world_axis_physical_types`
+            `ndcube.NDCube.wcs.world_axis_physical_types <astropy.wcs.wcsapi.BaseWCSWrapper>`
             of axes for which real world coordinates are desired.
             Not specifying axes inputs causes results for all axes to be returned.
         pixel_corners: `bool`, optional
@@ -139,14 +139,15 @@ class NDCubeABC(astropy.nddata.NDDataBase):
                                  ) -> Iterable[u.Quantity]:
         """
         Returns the world coordinate values of all pixels for desired axes.
-        In contrast to `axis_world_coords()`, this method returns `~astropy.units.Quantity` objects. which only
-        provide units rather than full coordinate metadata provided by high-level coordinate objects.
+        In contrast to :meth:`ndcube.NDCube.axis_world_coords`, this method returns
+        `~astropy.units.Quantity` objects. Which only provide units rather than full
+        coordinate metadata provided by high-level coordinate objects.
 
         Parameters
         ----------
         axes: `int` or `str`, or multiple `int` or `str`, optional
             Axis number in numpy ordering or unique substring of
-            `~ndcube.NDCube.wcs.world_axis_physical_types`
+            `ndcube.NDCube.wcs.world_axis_physical_types <astropy.wcs.wcsapi.BaseWCSWrapper>`
[…]
