+++ ndcube/visualization/__init__.py
@@ -1,4 +1,4 @@
 from .base import BasePlotter
 from .descriptor import PlotterDescriptor
 
-__all__ = ['BasePlotter']
+__all__ = ['BasePlotter', "PlotterDescriptor"]
