### 2023-06-21 #5296 @DanRyanIrish
@adonath I am happy to say that the associated ndcube ApJ paper has been accepted and we should receive a DOI soon.  Before that can happen though, it seems we need to provide ApJ with this paper's DOI.  Can you confirm that it will be:   https://doi.org/10.21105.joss.05296  ?

Can you also confirm what the month of this JOSS publication should be which it required for proper citing?
