### 2023-06-22 #5296 @adonath
Thanks a lot @DanRyanIrish, that's great news! My understand of the process is that the reference only goes in one direction:the ApJ DOI is cited in the JOSS paper, but no the other way around. However I'll tag the EiC @arfon to confirm.
