### 2023-06-22 #5296 @DanRyanIrish
If that is normally the case, perhaps it's required for this paper because I cite the JOSS paper in the ApJ one...?
