### 2023-06-23 #5296 @augustfly
@adonath I'm not certain what @arfon will recall, but we def pref bi-directional citing (AAS formally cites JOSS). We also banner the final article with the JOSS DOI. I believe the JOSS DOI is set the moment the review issue is begun (by the GitHub issue #, e.g., 5296).

The AAS DOI will be: 10.3847/1538-4357/ace0bd
