### 2023-06-26 #5296 @adonath
Thanks @augustfly for the clarification, this makes sense! I will request to maybe extend the JOSS guidelines for joint submissions (https://joss.readthedocs.io/en/latest/editing.html#handling-of-papers-published-together-with-aas-publishing) to mention and check the bi-directional citing.

@DanRyanIrish Yes, the JOSS DOI should be https://doi.org/10.21105/joss.05296. Please also update the meta data in the `paper.md` with the new AAS DOI. See https://joss.readthedocs.io/en/latest/editing.html#handling-of-papers-published-together-with-aas-publishing for instructions.

Here you can find an example BibTex from a JOSS paper:

```
@article{Riley2023,
	title        = {X-PSI: A Python package for neutron star X-ray pulse simulation and inference},
	author       = {Thomas E. Riley and Devarshi Choudhury and Tuomo Salmi and Serena Vinciguerra and Yves Kini and Bas Dorsman and Anna L. Watts and Daniela Huppenkothen and Sebastien Guillot},
	year         = 2023,
	journal      = {Journal of Open Source Software},
	publisher    = {The Open Journal},
	volume       = 8,
	number       = 82,
	pages        = 4977,
	doi          = {10.21105/joss.04977},
	url          = {https://doi.org/10.21105/joss.04977}
}
```
So this does not mention the month, however a it defines a volume and number. I'm not sure how this information is created. Can you please comment @arfon?
