+++ joss_paper/paper.bib
@@ -172,8 +172,9 @@ archivePrefix = {arXiv},
     author = {{Ryan}, Daniel and {Mumford}, Stuart and {Barnes}, Will T. and {Kumar Baruah}, Ankit and {Bhope}, Adwait and {Buchlin}, Eric and {Freij}, Nabil and {Ginsburg}, Adam and {Hayes}, Laura A. and {Homeier}, Derek and {Hughes}, J. Marcus and {Lowder}, Chris and {O'Steen}, Richard and {Pellorce}, Baptiste and {Robitaille}, Thomas and {Sharma}, Yash and {Shih}, Albert Y. and {Tollerud}, Erik and {West}, Matthew J.},
     title = "{A Unified Framework for Manipulating N-dimensional Astronomical Data and Coordinate
 Transformations in Python: The NDCube 2 & Astropy APE-14 WCS APIs}",
-    year = {submitted},
-    journal = {The Astrophysical Journal},
+    year = 2023,
+    journal = {ApJ},
+    doi = {10.3847/1538-4357/ace0bd},
 }
 
 @Article{numpy,
