### 2023-06-27 #5296 @DanRyanIrish
@adonath I've opened a [PR to the ndcube](https://github.com/sunpy/ndcube/pull/621/files) repo to update the bibtex reference.  However, I don't have all the info you asked for.

@augustfly Could you clarify what information we are able to supply at this time?  The link should take you directly to the bibtex reference being updated.  You can comment here or directly on that PR.  Thanks!
