### 2023-06-27 #5296 @arfon
> So this does not mention the month, however a it defines a volume and number. I'm not sure how this information is
created. Can you please comment @arfon?

It's generated by the month and year for when the JOSS paper is published. We're currently in volume 8 (and will be for all of this calendar year), and number (issue) 86 for this month. From July this year we'll be in issue 87, August 88 etc.

Also, the page number is the issue number here (5296).
