### 2023-06-27 #5296 @augustfly
> @adonath I've opened a [PR to the ndcube](https://github.com/sunpy/ndcube/pull/621/files) repo to update the bibtex reference. However, I don't have all the info you asked for.
>
> @augustfly Could you clarify what information we are able to supply at this time? The link should take you directly to the bibtex reference being updated. You can comment here or directly on that PR. Thanks!

@DanRyanIrish I think that JOSS handles this (citing/linking to the AAS DOI) through `paper.md` instead of the paper's bibliography.

The bibtex example @adonath gave is what we (at that AAS Journals) would need, but actually it's all fine from our end until you get our page proofs to double check things.

So I don't think you need full bibtex for the AAS DOI; only update `paper.md` with the DOI and journal title and then my read is that you're good to go.
