### 2023-06-30 #5296 @adonath
The one invalid DOI is expected...so looks good!
