### 2023-07-03 #5296 @adonath
@DanRyanIrish Here are some final minor comments on the paper from my side:
- Line 42: the reference for gWCS appears with a question marks and seems undefined
- Line 125: the reference to the ndcube docs seems undefined
- Line 151: the reference to the specutils code seems undefined
- Whenever you link to online documentation in the references I would recommend to not link to stable / latest, but provide the specific version you looked at instead. The latest / stable links will resolve to different versions, as time evolves.
- The reference to APE 14 currently goes to GitHub, I would recommend to use the  Zenodo DOI instead: https://zenodo.org/record/1188875
