### 2023-07-05 d37a60b
JOSS paper formatting corrections.
