+++ joss_paper/paper.bib
@@ -1,10 +1,15 @@
-@misc{ape14,
-  author = {{Robitaille}, Thomas and {Tollerud}, Erik and {Mumford}, Stuart and {Ginsburg}, Adam},
-  title = {APE-14: A shared Python interface for World Coordinate Systems},
-  year = {2018},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/astropy/astropy-APEs/blob/main/APE14.rst}
+@manual{ape14,
+  title        = {{Astropy Proposal for Enhancement 14: A shared
+                   Python interface for World Coordinate Systems (APE
+                   14)}},
+  author       = {Robitaille, Thomas and
+                  Tollerud, Erik and
+                  Mumford, Stuart and
+                  Ginsburg, Adam},
+  month        = feb,
+  year         = 2018,
+  doi          = {10.5281/zenodo.1188875},
+  url          = {https://doi.org/10.5281/zenodo.1188875}
 }
 
 @ARTICLE{astropy2013,
@@ -97,22 +102,38 @@ archivePrefix = {arXiv},
   url = {https://github.com/USNavalResearchLaboratory/eispac}
 }
 
-@misc{gwcs-docs,
-  author = {{gWCS Developers}},
-  title = {gWCS Documentation},
-  year = {2023},
-  publisher = {ReadTheDocs},
-  journal = {ReadTheDocs},
-  url = {https://gwcs.readthedocs.io/en/stable/}
-}
-
-@misc{gwcs-code,
-  author = {{gWCS Developers}},
-  title = {gWCS code-base},
-  year = {2023},
-  publisher = {GitHub},
-  journal = {GitHub Repository},
-  url = {https://github.com/spacetelescope/gwcs}
+@software{gwcs,
+  author       = {Nadia Dencheva and
+                  Stuart Mumford and
+                  Mihai Cara and
+                  Larry Bradley and
+                  perrygreenfield and
+                  Dan D'Avella and
+                  Brigitta Sipőcz and
+                  P. L. Lim and
+                  William Jamieson and
+                  Edward Slavich and
+                  Clare Shanahan and
+                  James Davies and
+                  Nicholas Earl and
+                  Zach Burnett and
+                  Bernie Simon and
+                  Erik Tollerud and
+                  Christoph Deil and
+                  Ole Streicher and
+                  Chris Simpson and
+                  David Stansby and
+                  Joseph Hunkeler and
+                  Michael Droettboom and
+                  Miguel de Val-Borro and
+                  Zane Geiger},
+  title        = {spacetelescope/gwcs: GWCS v 0.18.3},
+  month        = dec,
+  year         = 2022,
+  publisher    = {Zenodo},
+  version      = {0.18.3},
+  doi          = {10.5281/zenodo.7478201},
+  url          = {https://doi.org/10.5281/zenodo.7478201}
 }
 
 @misc{irispy-docs,
@@ -121,7 +142,7 @@ archivePrefix = {arXiv},
   year = {2023},
   publisher = {ReadTheDocs},
   journal = {ReadTheDocs},
-  url = {https://irispy-lmsal.readthedocs.io/en/stable/}
+  url = {https://irispy-lmsal.readthedocs.io/en/v0.1.5/}
 }
 
 @misc{irispy-code,
@@ -175,6 +196,15 @@ Transformations in Python: The NDCube 2 & Astropy APE-14 WCS APIs}",
     year = 2023,
     journal = {ApJ},
     doi = {10.3847/1538-4357/ace0bd},
+}
+
+@misc{ndcube-docs,
+  author = {{ndcube Developers}},
[…]
