### 2023-07-08 10847b4
Replaces internal use of numpy.product
with numpy.prod due to deprecation.
