+++ changelog/625.trivial.rst
@@ -1,1 +1,0 @@
-Remove internal use of numpy.product as its now deprecated and replace with numpy.prod as recommended.
