### 2023-07-08 #5296 @DanRyanIrish
Hi @adonath.  I am making a new ndcube release which includes the JOSS paper.  Regarding archiving the repo, I am not familiar with that process.  I obviously still want a version to be unarchived so future versions of ndcube can be work on.  Could you elaborate on how archiving is usually done for JOSS?
