### 2023-07-10 #5296 @adonath
@DanRyanIrish Most people choose Zenodo (https://zenodo.org) as an archive for the source code. But there is also e.g. the https://ascl.net or https://figshare.com. It is mostly your choice, the important point is to get a DOI, we can link in the meta data. I would recommend to just upload to Zenodo and report the ODI here. Thanks!
