### 2023-07-10 #5296 @timj
You can configure GitHub to automatically upload releases to Zenodo when you make a formal GitHub Release.
