### 2023-07-10 #5296 @adonath
Instructions are here: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content#issuing-a-persistent-identifier-for-your-repository-with-zenodo

But I don't think you can do this retrospectively, for past release you probably have to do a manual upload.
