### 2023-07-15 a043622
Remove commented out test file
