+++ ndcube/tests/test_sequence_plotting.py
@@ -1,777 +1,0 @@
-import copy
-import datetime
-
-import astropy.units as u
-import matplotlib
-import numpy as np
-import pytest
-
-import ndcube.mixins.sequence_plotting
-from ndcube import NDCube, NDCubeSequence
-from ndcube.utils.wcs import WCS
-
-# sample data for tests
-# TODO: use a fixture reading from a test file. file TBD.
-data = np.array([[[1, 2, 3, 4], [2, 4, 5, 3], [0, -1, 2, 3]],
-                 [[2, 4, 5, 1], [10, 5, 2, 2], [10, 3, 3, 0]]])
-
-data2 = np.array([[[11, 22, 33, 44], [22, 44, 55, 33], [0, -1, 22, 33]],
-                  [[22, 44, 55, 11], [10, 55, 22, 22], [10, 33, 33, 0]]])
-
-ht = {'CTYPE3': 'HPLT-TAN', 'CUNIT3': 'deg', 'CDELT3': 0.5, 'CRPIX3': 0, 'CRVAL3': 0, 'NAXIS3': 2,
-      'CTYPE2': 'WAVE    ', 'CUNIT2': 'Angstrom', 'CDELT2': 0.2, 'CRPIX2': 0, 'CRVAL2': 0,
-      'NAXIS2': 3,
-      'CTYPE1': 'TIME    ', 'CUNIT1': 'min', 'CDELT1': 0.4, 'CRPIX1': 0, 'CRVAL1': 0, 'NAXIS1': 4}
-wt = WCS(header=ht, naxis=3)
-
-hm = {
-    'CTYPE1': 'WAVE    ', 'CUNIT1': 'Angstrom', 'CDELT1': 0.2, 'CRPIX1': 0, 'CRVAL1': 10,
-    'NAXIS1': 4,
-    'CTYPE2': 'HPLT-TAN', 'CUNIT2': 'deg', 'CDELT2': 0.5, 'CRPIX2': 2, 'CRVAL2': 0.5, 'NAXIS2': 3,
-    'CTYPE3': 'HPLN-TAN', 'CUNIT3': 'deg', 'CDELT3': 0.4, 'CRPIX3': 2, 'CRVAL3': 1, 'NAXIS3': 2}
-wm = WCS(header=hm, naxis=3)
-
-
-cube1 = NDCube(
-    data, wt, missing_axes=[False, False, False, True],
-    extra_coords=[
-        ('pix', 0, u.Quantity(range(data.shape[0]), unit=u.pix)),
-        ('hi', 1, u.Quantity(range(data.shape[1]), unit=u.s)),
-        ('distance', None, u.Quantity(0, unit=u.cm)),
-        ('time', None, datetime.datetime(2000, 1, 1, 0, 0))])
-
-cube1_with_unit = NDCube(
-    data, wt, missing_axes=[False, False, False, True],
-    unit=u.km,
-    extra_coords=[
-        ('pix', 0, u.Quantity(range(data.shape[0]), unit=u.pix)),
-        ('hi', 1, u.Quantity(range(data.shape[1]), unit=u.s)),
-        ('distance', None, u.Quantity(0, unit=u.cm)),
-        ('time', None, datetime.datetime(2000, 1, 1, 0, 0))])
-
-cube1_with_mask = NDCube(
-    data, wt, missing_axes=[False, False, False, True],
-    mask=np.zeros_like(data, dtype=bool),
-    extra_coords=[
-        ('pix', 0, u.Quantity(range(data.shape[0]), unit=u.pix)),
-        ('hi', 1, u.Quantity(range(data.shape[1]), unit=u.s)),
-        ('distance', None, u.Quantity(0, unit=u.cm)),
[…]
