+++ setup.cfg
@@ -173,4 +173,5 @@ ignore-words-list =
     observ,
     ot,
     te,
-    upto
+    upto,
+    aas
