### 2023-07-25 88d6953
edits to manuscript for JOSS review
