+++ joss_paper/paper.md
@@ -100,10 +100,10 @@ analogous to how users handle coordinate-agnostic data with arrays.
 ndcube requires that coordinate transformations be expressed via the World
 Coordinate System (WCS), a coordinate framework commonly used throughout
 astronomy.
-The WCS framework has multiple implementations, e.g. FITS-WCS, gWCS, etc.,
+The WCS framework has multiple implementations (e.g. FITS-WCS, gWCS, and others),
 each with a different incompatible API, which makes workflows and derived tools
 non-transferable between implementations.
-ndcube overcomes this by leveraging AstroPy's WCS API [APE-14; @ape14]
+ndcube overcomes this by leveraging Astropy's WCS API [APE-14, @ape14]
 which can be wrapped around any underlying WCS implementation.
 This enables ndcube to use the same API to interact with any set of WCS transformations.
 ndcube's data-WCS coupling allows users to analyze their data more easily and
@@ -116,7 +116,7 @@ N-dimensional data sets are common in all areas of science and beyond.
 For example, a series of images taken sequentially with a CCD camera can be stored
 as a single 3-D array with two spatial axes and one temporal axis.
 In astronomy, the most commonly used framework for translating between array element
-indices and the location, time, etc. in the Universe being observed is the World
+indices, and the location or time in the Universe being observed is the World
 Coordinate System (WCS).
 WCS's ability to handle many different physical types of coordinates (e.g. spatial,
 temporal, spectral, etc.) and their projections onto a data array (e.g. right ascension
@@ -125,14 +125,14 @@ standardized and powerful way to relate array axes to the physical coordinate ty
 they represent.
 
 There are mature Python packages for handling N-D array operations --
-e.g. numpy [@numpy], dask [@dask], etc. -- and others for supporting WCS coordinate
-transformations -- e.g. astropy [@astropy2013; @astropy2018; @astropy2022], gWCS [@gwcs].
+for example, NumPy [@numpy], and Dask [@dask] -- and others for supporting WCS coordinate
+transformations -- for example, Astropy [@astropy2013; @astropy2018; @astropy2022], and gWCS [@gwcs].
 However, none treat data and coordinates in a combined, self-consistent way.
-The closest alternative to ndcube is xarray [@xarray].
-However xarray has been developed for the requirements and conventions of the
+The closest alternative to ndcube is Xarray [@xarray].
+However Xarray has been developed for the requirements and conventions of the
 geosciences which, although similar to those of astronomy in concept, are sufficiently
 different in construction to cause significant friction.
-Crucially, xarray does not currently support WCS coordinate transformations.
+Crucially, Xarray does not currently support WCS coordinate transformations.
 Tools that do support WCS-based coordinate-aware data analysis, such as the SunPy
 [@sunpy] Map class for 2-D images of the Sun, tend to have APIs specific to particular
 combinations of dimensions, physical types, coordinate systems and WCS implementations.
@@ -143,8 +143,8 @@ astronomy.
 
 ndcube overcomes these challenges via its design policy that all functionalities and
 APIs must be agnostic to the number of dimensions and coordinate types they represent.
-Moreover, ndcube's employment of the AstroPy WCS API makes it agnostic to the
-underlying WCS implementation, e.g. FITS-WCS, gWCS, etc.
+Moreover, ndcube's employment of the Astropy WCS API makes it agnostic to the
+underlying WCS implementation.
 
 
 # The Role of ndcube and its Features
@@ -154,7 +154,7 @@ First, it formalizes the NDCube 2 API in software via its abstract base classes
 (ABCs), `NDCubeABC`, `ExtraCoordsABC` and `GlobalCoordsABC`.
[…]
