### 2023-07-26 #5296 @DanRyanIrish
Hi @dfm.  Thanks for you PR.  THe changes look good to me and a member of the ndcube team has already merged them.  Please proceed with the processing as you see fit.  Many thanks!
