### 2023-07-26 #5296 @dfm
@DanRyanIrish — Thanks! I have one other thing I need from you actually, sorry for not catching it sooner! Can you update the metadata for the Zenodo archive (you should see an edit button on the record page if you're logged in) so that the title and author list are identical to the paper?
