### 2023-07-26 #5296 @dfm
@DanRyanIrish — thanks for checking! Whoever set up the GitHub integration should have these permissions. Can you ask the team? Alternatively, you could manually create an archive that you own and edit its metadata? Thanks!!
