### 2023-08-28 aa9712a
Backport PR #632: DOC: Remove broken link from old mailing list
