+++ setup.cfg
@@ -27,7 +27,6 @@ classifiers =
   Operating System :: OS Independent
   Programming Language :: Python
   Programming Language :: Python :: 3
-  Programming Language :: Python :: 3.8
   Programming Language :: Python :: 3.9
   Programming Language :: Python :: 3.10
   Programming Language :: Python :: 3.11
@@ -37,11 +36,11 @@ classifiers =
 zip_safe = False
 packages = find:
 include_package_data = True
-python_requires = >=3.8
+python_requires = >=3.9
 install_requires =
-  astropy>=4.2
-  gwcs>=0.15
-  numpy>1.17
+  astropy>=5.0.6,!=5.1.0
+  gwcs>=0.18
+  numpy>=1.21.0
 
 [options.extras_require]
 tests =
@@ -50,7 +49,7 @@ tests =
     pytest-astropy
     pytest-mpl>=0.12
     scipy
-    sunpy>=4.0.0
+    sunpy>=5.0.0
 docs =
     matplotlib
     pytest-doctestplus>=0.9.0
@@ -60,9 +59,9 @@ docs =
     sphinx-gallery
     sphinxext-opengraph
     sunpy-sphinx-theme
-    sunpy>=4.0.0
+    sunpy>=5.0.0
 plotting =
-    matplotlib>=3.2
+    matplotlib>=3.5.0
     mpl_animators>=1.0
 reproject =
     reproject>=0.7.1
