+++ tox.ini
@@ -1,14 +1,9 @@
 [tox]
+min_version = 4.0
 envlist =
-    py{38,39,310}{,-devdeps,-online,-figure,-conda,-oldestdeps}
+    py{39,310,311}{,-devdeps,-online,-figure,-conda,-oldestdeps}
     build_docs
     codestyle
-requires =
-    setuptools >=56, !=61.0.0
-    pip >= 19.3.1
-    tox < 4
-    tox-pypi-filter >= 0.12
-isolated_build = true
 
 [testenv]
 pypi_filter = https://raw.githubusercontent.com/sunpy/sunpy/main/.test_package_pins.txt
