+++ tox.ini
@@ -76,6 +76,7 @@ commands =
     figure: {env:PYTEST_COMMAND} -m "mpl_image_compare" --mpl --remote-data=any --mpl-generate-summary=html --mpl-baseline-path=https://raw.githubusercontent.com/sunpy/sunpy-figure-tests/ndcube-main/figures/{envname}/ {posargs}
 install_command =
     devdeps: python -m pip install --extra-index-url=https://pypi.anaconda.org/astropy/simple {opts} {packages}
+    !devdeps: python -m pip install {packages}
 
 [testenv:build_docs]
 changedir = docs
@@ -99,7 +100,7 @@ commands =
     pre-commit run --color always --all-files --show-diff-on-failure
 
 # This env requires tox-conda.
-[testenv:py{38,39,310,311}-conda]
+[testenv:py{39,310,311}-conda]
 pypi_filter =
 extras =
 deps =
