+++ tox.ini
@@ -52,17 +52,11 @@ deps =
     oldestdeps: scipy<1.8.0
     oldestdeps: dask<2021.5.0
     # Figure tests need a tightly controlled environment
-    # These are the oldestdeps which have py310 wheels as running multiple
-    # versions of Python on circleci is a little annoying.
-    # Newer versions are covered by devdeps
-    figure-!devdeps: matplotlib==3.5.3
-    figure-!devdeps: astropy==5.1.0
+    figure-!devdeps: matplotlib==3.7.2
+    figure-!devdeps: astropy==5.3.3
     figure-!devdeps: mpl-animators==1.0.0
     figure-!devdeps: scipy
     figure-!devdeps: dask
-    figure-!devdeps: numpy<1.24
-    # For some reason gwcs 0.18.2 bumped astropy to 5.1 min
-    figure-!devdeps: gwcs<0.18.2
 extras =
     all
     tests
