+++ tox.ini
@@ -23,7 +23,7 @@ setenv =
     COLUMNS = 180
     PYTEST_COMMAND = pytest -vvv -s -raR --pyargs ndcube --cov-report=xml --cov=ndcube --cov-config={toxinidir}/setup.cfg {toxinidir}/docs
     PARFIVE_HIDE_PROGESS = True
-    devdeps: PIP_EXTRA_INDEX_URL = https://pypi.anaconda.org/astropy/simple https://pypi.anaconda.org/scipy-wheels-nightly/simple
+    devdeps: PIP_EXTRA_INDEX_URL = https://pypi.anaconda.org/astropy/simple https://pypi.anaconda.org/scientific-python-nightly-wheels/simple
 passenv =
     HTTP_PROXY
     HTTPS_PROXY
@@ -68,9 +68,6 @@ commands =
     figure: /bin/bash -c "cat ./figure_test_images/figure_version_info.txt"
     figure: python -c "import ndcube.tests.helpers as h; print(h.get_hash_library_name())"
     figure: {env:PYTEST_COMMAND} -m "mpl_image_compare" --mpl --remote-data=any --mpl-generate-summary=html --mpl-baseline-path=https://raw.githubusercontent.com/sunpy/sunpy-figure-tests/ndcube-main/figures/{envname}/ {posargs}
-install_command =
-    devdeps: python -m pip install --extra-index-url=https://pypi.anaconda.org/astropy/simple {opts} {packages}
-    !devdeps: python -m pip install {packages}
 
 [testenv:build_docs]
 changedir = docs
