### 2023-09-22 #5296 @DanRyanIrish
Hi @arfon @dfm .  Thanks for your patience.  We have not been able to establish who has edit privileges on the zenodo publication and so have not been able to adjust the author list.  I think the best thing to do at this stage is to simply keep the author lists different between zenodo and JOSS as they are now.
