### 2023-09-22 #5296 @adonath
@DanRyanIrish Which one is the "correct" author list? The paper or Zenodo? In any case I can really recommend adding a `CITATION.cff` file to the repo. This way you have more control on the author list and can also add more meta data such as orcids etc., the `CITATION.cff` file is also supported by Zenodo.
