### 2023-09-22 #5296 @dfm
@DanRyanIrish — Thanks for digging into this! It is OK to have different metadata in the two places; our preference is to have them consistent, but it's not the end of the world! @adonath's suggestion is a good one for the future! For now I'm going to get this review wrapped up.
